moving some more utility classes to separate folder
