+++ swiftwinrt/code_writers/enum_writers.cpp
@@ -1,4 +1,4 @@
-#include "metadata_cache.h"
+#include "utility/metadata_cache.h"
 #include "common_writers.h"
 #include "type_writers.h"
 #include "can_write.h"
