+++ swiftwinrt/pch.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "cmd_reader.h"
+#include "utility/cmd_reader.h"
 #include <winmd_reader.h>
-#include "task_group.h"
-#include "text_writer.h"
+#include "utility/task_group.h"
+#include "utility/text_writer.h"
