+++ swiftwinrt/utility/attributes.h
