+++ swiftwinrt/utility/cmd_reader.h
