+++ swiftwinrt/utility/metadata_cache.h
