+++ swiftwinrt/utility/metadata_filter.h
