+++ swiftwinrt/utility/settings.h
