+++ swiftwinrt/utility/sha1.h
