+++ swiftwinrt/utility/task_group.h
