+++ swiftwinrt/utility/text_writer.h
