+++ swiftwinrt/utility/versioning.h
